refactor: rename module to `typer_ext`
